chore: moved video utilities to common
