+++ components/src/dynamo/common/utils/video_utils.py
@@ -1,9 +1,10 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Video encoding utilities for TensorRT-LLM video diffusion.
+"""Shared video encoding utilities.
 
-This module provides utilities for encoding numpy video frames to MP4 format.
+This module provides utilities for converting PIL images to numpy arrays
+and encoding numpy video frames to MP4 format (on disk or in memory).
 """
 
 import io
@@ -15,26 +16,58 @@ import numpy as np
 logger = logging.getLogger(__name__)
 
 
+def frames_to_numpy(images: list) -> np.ndarray:
+    """Convert a list of PIL Images to a numpy array suitable for video encoding.
+
+    Args:
+        images: List of PIL Image objects (video frames).
+
+    Returns:
+        Numpy array of shape ``(num_frames, height, width, 3)`` with dtype
+        ``uint8`` and values in ``[0, 255]``.
+
+    Raises:
+        ValueError: If no images are provided or images have inconsistent sizes.
+    """
+    if not images:
+        raise ValueError("No images provided for video encoding")
+
+    frames = []
+    for img in images:
+        arr = np.array(img.convert("RGB"))
+        frames.append(arr)
+
+    # Validate consistent sizes
+    shapes = {f.shape for f in frames}
+    if len(shapes) > 1:
+        raise ValueError(
+            f"Inconsistent frame sizes detected: {shapes}. "
+            "All frames must have the same dimensions."
+        )
+
+    return np.stack(frames, axis=0)
+
+
 def encode_to_mp4(
     frames: np.ndarray,
     output_dir: str,
     request_id: str,
     fps: int = 16,
 ) -> str:
-    """Encode numpy frames to MP4 file.
+    """Encode numpy frames to an MP4 file on disk.
 
     Args:
-        frames: Video frames as numpy array of shape (num_frames, height, width, 3)
-            with uint8 values 0-255.
+        frames: Video frames as numpy array of shape
+            ``(num_frames, height, width, 3)`` with uint8 values ``[0, 255]``.
         output_dir: Directory to save the output video.
-        request_id: Unique identifier for the request (used in filename).
+        request_id: Unique request identifier (used in filename).
         fps: Frames per second for the output video.
 
     Returns:
-        Path to the saved MP4 file.
+        Absolute path to the saved MP4 file.
 
     Raises:
-        ImportError: If imageio is not available.
+        ImportError: If ``imageio`` is not available.
         RuntimeError: If encoding fails.
     """
     try:
@@ -42,20 +75,18 @@ def encode_to_mp4(
     except ImportError:
         try:
             import imageio as iio
-        except ImportError:
+        except ImportError as err:
             raise ImportError(
                 "imageio is required for video encoding. "
                 "Install with: pip install imageio[ffmpeg]"
-            )
+            ) from err
 
-    # Ensure output directory exists
     os.makedirs(output_dir, exist_ok=True)
     output_path = os.path.join(output_dir, f"{request_id}.mp4")
 
     logger.info(f"Encoding {len(frames)} frames to {output_path} at {fps} fps")
 
     try:
-        # Use imageio to write MP4
         # imageio.v3 API
         if hasattr(iio, "imwrite"):
             iio.imwrite(output_path, frames, fps=fps, codec="libx264")
@@ -72,7 +103,7 @@ def encode_to_mp4(
         return output_path
 
     except Exception as e:
-        logger.error(f"Failed to encode video: {e}")
+        logger.exception(f"Failed to encode video: {e}")
         raise RuntimeError(f"Video encoding failed: {e}") from e
 
 
@@ -80,18 +111,18 @@ def encode_to_mp4_bytes(
     frames: np.ndarray,
     fps: int = 16,
 ) -> bytes:
-    """Encode numpy frames to MP4 bytes (in-memory).
+    """Encode numpy frames to MP4 bytes in memory.
 
     Args:
-        frames: Video frames as numpy array of shape (num_frames, height, width, 3)
-            with uint8 values 0-255.
+        frames: Video frames as numpy array of shape
+            ``(num_frames, height, width, 3)`` with uint8 values ``[0, 255]``.
         fps: Frames per second for the output video.
 
     Returns:
-        MP4 video as bytes.
+        MP4 video as raw bytes.
 
     Raises:
-        ImportError: If imageio is not available.
+        ImportError: If ``imageio`` is not available.
         RuntimeError: If encoding fails.
     """
     try:
@@ -99,19 +130,17 @@ def encode_to_mp4_bytes(
     except ImportError:
         try:
             import imageio as iio
-        except ImportError:
+        except ImportError as err:
             raise ImportError(
                 "imageio is required for video encoding. "
                 "Install with: pip install imageio[ffmpeg]"
-            )
+            ) from err
 
     logger.info(f"Encoding {len(frames)} frames to bytes at {fps} fps")
 
     try:
-        # Use in-memory buffer
         buffer = io.BytesIO()
 
-        # imageio can write to BytesIO with format hint
         if hasattr(iio, "imwrite"):
             # v3 API - write to buffer
             iio.imwrite(buffer, frames, extension=".mp4", fps=fps, codec="libx264")
@@ -131,5 +160,5 @@ def encode_to_mp4_bytes(
         return video_bytes
 
     except Exception as e:
-        logger.error(f"Failed to encode video to bytes: {e}")
+        logger.exception(f"Failed to encode video to bytes: {e}")
         raise RuntimeError(f"Video encoding to bytes failed: {e}") from e
